[SCRIPTS] Handle already fixed uids

+++ bin/upgrade/dev/scripts/fix_nxret.js
@@ -17,6 +17,7 @@ class WriteHandler {
         this.nDocs = 0;
         this.nInsertRequests = 0;
         this.nRemoveRequests = 0;
+        this.nAlreadyFixed = 0;
     }
 
     flush() {
@@ -31,30 +32,38 @@ class WriteHandler {
         var ranking = this.candidates.map(function(doc){
             var str = JSON.stringify(doc);
             return {
+                "fs": parseInt(doc.fs),
+                "len": str.length,
+                "fixed": doc._id === doc.uid ? 1 : 0,
                 "doc": doc, 
                 "removeId": doc._id,
-                "fs": parseInt(doc.fs),
-                "str": str,
-                "len": str.length
+                "str": str
             }
         });
         ranking.sort(function(a, b) {
-            return a.fs - b.fs || b.len - a.len;
+            return b.fixed - a.fixed || a.fs - b.fs || b.len - a.len;
         });
         var winningItem = ranking[0];
         winningItem.doc._id = winningItem.doc.uid;
         var self = this;
         
-        requests.push({
-            insertOne: { "document": winningItem.doc }
-        });
-        self.nInsertRequests++;
+        if (winningItem.fixed !== 1) {
+            requests.push({
+                insertOne: { "document": winningItem.doc }
+            });
+            self.nInsertRequests++;
+        }
+        else {
+            self.nAlreadyFixed++;
+        }
 
         ranking.forEach(function(item){
-            requests.push({
-                deleteOne: { "filter": { _id: item.removeId }}
-            });
-            self.nRemoveRequests++;
+            if (item.fixed !== 1) {
+                requests.push({
+                    deleteOne: { "filter": { _id: item.removeId }}
+                });
+                self.nRemoveRequests++;
+            }
         });
         return requests;
     }
@@ -109,7 +118,7 @@ countlyDb.collection('apps').find({}).toArray(function(appsErr, apps) {
 
     function upgrade(app, done) {
         printMessage("log", "(" + app.name + ") Fixing...");
-        var cursor = countlyDb.collection('app_nxret' + app._id).find({$expr: {$ne: ["$_id", "$uid"] } }).sort({uid: 1});
+        var cursor = countlyDb.collection('app_nxret' + app._id).find({}).sort({uid: 1});
         var requests = [];
         var nScanned = 0;
         var nSkipped = 0;
@@ -120,10 +129,6 @@ countlyDb.collection('apps').find({}).toArray(function(appsErr, apps) {
         cursor.forEach(function(nxret) {
             if (nxret.uid === null || nxret.uid === undefined) {
                 printMessage("log", "(" + app.name + ") Skipping a doc with empty uid");
-                nSkipped++;
-            }
-            else if (nxret._id === nxret.uid) {
-                printMessage("log", "(" + app.name + ") Skipping a doc whose _id = uid");
                 nSkipped++;
             }
             else {
@@ -200,9 +205,10 @@ countlyDb.collection('apps').find({}).toArray(function(appsErr, apps) {
             }
             else {
                 printMessage("log", "(" + app.name + ")", "scanned =", nScanned, "/", "skipped =", nSkipped);
-                printMessage("log", "(" + app.name + ")", "uids with dup =", writeHandler.nDocsWithDups, "/", "total unique uids =", writeHandler.nDocs);
                 if (nScanned > 0) {
-                    printMessage("log", "(" + app.name + ")", "insert reqs=", writeHandler.nInsertRequests, "/", "remove reqs =", writeHandler.nRemoveRequests);
+                    printMessage("log", "(" + app.name + ")", "uids with dup =", writeHandler.nDocsWithDups, "/", "total unique uids =", writeHandler.nDocs);
+                    printMessage("log", "(" + app.name + ")", "uids already fixed =", writeHandler.nAlreadyFixed);
+                    printMessage("log", "(" + app.name + ")", "insert reqs =", writeHandler.nInsertRequests, "/", "remove reqs =", writeHandler.nRemoveRequests);
                     printMessage("log", "(" + app.name + ")", "inserted =", nInserted, "/", "removed =", nRemoved);
                 }
                 if (writeHandler.nInsertRequests === nInserted && writeHandler.nRemoveRequests === nRemoved) {
